+++ client/src/main/java/com/fastcomments/model/FComment.java
@@ -400,7 +400,7 @@ public class FComment {
   public static final String SERIALIZED_NAME_VIEW_COUNT = "viewCount";
   @SerializedName(SERIALIZED_NAME_VIEW_COUNT)
   @javax.annotation.Nullable
-  private Double viewCount;
+  private Long viewCount;
 
   public static final String SERIALIZED_NAME_REQUIRES_VERIFICATION = "requiresVerification";
   @SerializedName(SERIALIZED_NAME_REQUIRES_VERIFICATION)
@@ -1763,7 +1763,7 @@ public class FComment {
   }
 
 
-  public FComment viewCount(@javax.annotation.Nullable Double viewCount) {
+  public FComment viewCount(@javax.annotation.Nullable Long viewCount) {
     this.viewCount = viewCount;
     return this;
   }
@@ -1773,11 +1773,11 @@ public class FComment {
    * @return viewCount
    */
   @javax.annotation.Nullable
-  public Double getViewCount() {
+  public Long getViewCount() {
     return viewCount;
   }
 
-  public void setViewCount(@javax.annotation.Nullable Double viewCount) {
+  public void setViewCount(@javax.annotation.Nullable Long viewCount) {
     this.viewCount = viewCount;
   }
